created useragent class
